Ajout des tests unitaires avec GTest
Mis en place des tests unitaires disponible avec la commande ```make t```

+++ src/mainDebug.cpp
@@ -6,7 +6,7 @@
 int main() {
     Node n(2);
     std::cout << "reseau :" << n.getWeight(0) << " / " << n.getWeight(1) << "/ b :" << n.getBias() << std::endl;
-    std::cout << "output (0,0) : " << n.processOutputs(std::vector<int> {0,0}) << std::endl;
-    std::cout << "output (1,1) : " << n.processOutputs(std::vector<int> {1,1}) << std::endl;
+    std::cout << "output (0,0) : " << n.processOutputs(std::vector<double> {0,0}) << std::endl;
+    std::cout << "output (1,1) : " << n.processOutputs(std::vector<double> {1,1}) << std::endl;
     std::cout << "TESTER" << std::endl;
 }
